+++ CHANGES.rst
@@ -1,10 +1,17 @@
 3.1.1 (unreleased)
 ==================
 
+- Automatically remove old blobs backups that have no corresponding filestorage backup.
+  We compare the timestamp of the oldest filestorage backup with the timestamps of the
+  blob backups.  This can be the name, if you use ``blob_timestamps = true``,
+  or the modification date of the blob backup.
+  This means that the ``keep_blob_days`` option is ignored, unless you use ``only_blobs = true``.
+  [maurits]
+
 - When backing up a blobstorage, use the timestamp of the latest filestorage backup.
   If a blob backup with that name is already there, then there were no database changes,
   so we do not make a backup.
-  This is only done when you use the new `blob_timestamps = true` option.
+  This is only done when you use the new ``blob_timestamps = true`` option.
   [maurits]
 
 - When restoring to a specific date, find the first blob backup at or before